Fix issue that JpaTicketRegist can not  presist OAuth Tickets
