+++ cas-server-support-oauth-core/src/main/java/org/apereo/cas/ticket/accesstoken/AccessTokenImpl.java
@@ -1,34 +1,34 @@
-package org.apereo.cas.support.oauth.ticket.refreshtoken;
+package org.apereo.cas.ticket.accesstoken;
 
 import org.apereo.cas.authentication.Authentication;
 import org.apereo.cas.authentication.principal.Service;
-import org.apereo.cas.support.oauth.ticket.code.OAuthCodeImpl;
+import org.apereo.cas.ticket.code.OAuthCodeImpl;
 import org.apereo.cas.ticket.ExpirationPolicy;
 
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
 
 /**
- * An OAuth refresh token implementation.
+ * An OAuth access token implementation.
  *
  * @author Jerome Leleu
  * @since 5.0.0
  */
 @Entity
-@DiscriminatorValue(RefreshToken.PREFIX)
-public class RefreshTokenImpl extends OAuthCodeImpl implements RefreshToken {
+@DiscriminatorValue(AccessToken.PREFIX)
+public class AccessTokenImpl extends OAuthCodeImpl implements AccessToken {
 
-    private static final long serialVersionUID = -3544459978950667758L;
+    private static final long serialVersionUID = 2339545346159721563L;
 
     /**
-     * Instantiates a new OAuth refresh token.
+     * Instantiates a new OAuth access token.
      */
-    public RefreshTokenImpl() {
+    public AccessTokenImpl() {
         // exists for JPA purposes
     }
 
     /**
-     * Constructs a new refresh token with unique id for a service and authentication.
+     * Constructs a new access token with unique id for a service and authentication.
      *
      * @param id the unique identifier for the ticket.
      * @param service the service this ticket is for.
@@ -36,9 +36,8 @@ public class RefreshTokenImpl extends OAuthCodeImpl implements RefreshToken {
      * @param expirationPolicy the expiration policy.
      * @throws IllegalArgumentException if the service or authentication are null.
      */
-    public RefreshTokenImpl(final String id,
-                            final Service service, final Authentication authentication,
-                            final ExpirationPolicy expirationPolicy) {
+    public AccessTokenImpl(final String id, final Service service, final Authentication authentication,
+                           final ExpirationPolicy expirationPolicy) {
         super(id, service, authentication, expirationPolicy);
     }
 }
