+++ cas-server-support-oauth/src/main/java/org/apereo/cas/config/CasOAuthConfiguration.java
@@ -14,15 +14,15 @@ import org.apereo.cas.support.oauth.OAuthConstants;
 import org.apereo.cas.support.oauth.authenticator.OAuthClientAuthenticator;
 import org.apereo.cas.support.oauth.authenticator.OAuthUserAuthenticator;
 import org.apereo.cas.support.oauth.services.OAuthCallbackAuthorizeService;
-import org.apereo.cas.support.oauth.ticket.accesstoken.AccessTokenFactory;
-import org.apereo.cas.support.oauth.ticket.accesstoken.DefaultAccessTokenFactory;
-import org.apereo.cas.support.oauth.ticket.accesstoken.OAuthAccessTokenExpirationPolicy;
-import org.apereo.cas.support.oauth.ticket.code.DefaultOAuthCodeFactory;
-import org.apereo.cas.support.oauth.ticket.code.OAuthCodeExpirationPolicy;
-import org.apereo.cas.support.oauth.ticket.code.OAuthCodeFactory;
-import org.apereo.cas.support.oauth.ticket.refreshtoken.DefaultRefreshTokenFactory;
-import org.apereo.cas.support.oauth.ticket.refreshtoken.OAuthRefreshTokenExpirationPolicy;
-import org.apereo.cas.support.oauth.ticket.refreshtoken.RefreshTokenFactory;
+import org.apereo.cas.ticket.accesstoken.AccessTokenFactory;
+import org.apereo.cas.ticket.accesstoken.DefaultAccessTokenFactory;
+import org.apereo.cas.ticket.accesstoken.OAuthAccessTokenExpirationPolicy;
+import org.apereo.cas.ticket.code.DefaultOAuthCodeFactory;
+import org.apereo.cas.ticket.code.OAuthCodeExpirationPolicy;
+import org.apereo.cas.ticket.code.OAuthCodeFactory;
+import org.apereo.cas.ticket.refreshtoken.DefaultRefreshTokenFactory;
+import org.apereo.cas.ticket.refreshtoken.OAuthRefreshTokenExpirationPolicy;
+import org.apereo.cas.ticket.refreshtoken.RefreshTokenFactory;
 import org.apereo.cas.support.oauth.validator.OAuth20ValidationServiceSelectionStrategy;
 import org.apereo.cas.support.oauth.validator.OAuthValidator;
 import org.apereo.cas.support.oauth.web.AccessTokenResponseGenerator;
