+++ cas-server-support-oauth/src/main/java/org/apereo/cas/support/oauth/web/OAuth20AccessTokenController.java
@@ -9,11 +9,11 @@ import org.apereo.cas.support.oauth.OAuthConstants;
 import org.apereo.cas.support.oauth.profile.OAuthClientProfile;
 import org.apereo.cas.support.oauth.profile.OAuthUserProfile;
 import org.apereo.cas.support.oauth.services.OAuthRegisteredService;
-import org.apereo.cas.support.oauth.ticket.OAuthToken;
-import org.apereo.cas.support.oauth.ticket.accesstoken.AccessToken;
-import org.apereo.cas.support.oauth.ticket.code.OAuthCode;
-import org.apereo.cas.support.oauth.ticket.refreshtoken.RefreshToken;
-import org.apereo.cas.support.oauth.ticket.refreshtoken.RefreshTokenFactory;
+import org.apereo.cas.ticket.OAuthToken;
+import org.apereo.cas.ticket.accesstoken.AccessToken;
+import org.apereo.cas.ticket.code.OAuthCode;
+import org.apereo.cas.ticket.refreshtoken.RefreshToken;
+import org.apereo.cas.ticket.refreshtoken.RefreshTokenFactory;
 import org.apereo.cas.support.oauth.util.OAuthUtils;
 import org.pac4j.core.context.J2EContext;
 import org.pac4j.core.profile.ProfileManager;
