+++ cas-server-support-oauth/src/main/java/org/apereo/cas/support/oauth/web/OAuth20AccessTokenResponseGenerator.java
@@ -7,8 +7,8 @@ import org.apache.http.HttpStatus;
 import org.apereo.cas.authentication.principal.Service;
 import org.apereo.cas.support.oauth.OAuthConstants;
 import org.apereo.cas.support.oauth.services.OAuthRegisteredService;
-import org.apereo.cas.support.oauth.ticket.accesstoken.AccessToken;
-import org.apereo.cas.support.oauth.ticket.refreshtoken.RefreshToken;
+import org.apereo.cas.ticket.accesstoken.AccessToken;
+import org.apereo.cas.ticket.refreshtoken.RefreshToken;
 import org.apereo.cas.support.oauth.util.OAuthUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
