+++ cas-server-support-oauth/src/main/java/org/apereo/cas/ticket/accesstoken/AccessTokenFactory.java
@@ -1,23 +1,23 @@
-package org.apereo.cas.support.oauth.ticket.code;
+package org.apereo.cas.ticket.accesstoken;
 
 import org.apereo.cas.authentication.Authentication;
 import org.apereo.cas.authentication.principal.Service;
 import org.apereo.cas.ticket.TicketFactory;
 
 /**
- * Factory to create OAuth codes.
+ * Factory to create OAuth access tokens.
  *
  * @author Jerome Leleu
  * @since 5.0.0
  */
-public interface OAuthCodeFactory extends TicketFactory {
+public interface AccessTokenFactory extends TicketFactory {
 
     /**
-     * Create an OAuth code.
+     * Create an access token.
      *
      * @param service the service
      * @param authentication the authentication
-     * @return the OAuth code
+     * @return the access token
      */
-    OAuthCode create(Service service, Authentication authentication);
+    AccessToken create(Service service, Authentication authentication);
 }
