+++ cas-server-support-oauth/src/main/java/org/apereo/cas/ticket/code/OAuthCodeExpirationPolicy.java
@@ -1,4 +1,4 @@
-package org.apereo.cas.support.oauth.ticket.code;
+package org.apereo.cas.ticket.code;
 
 import org.apereo.cas.ticket.support.MultiTimeUseOrTimeoutExpirationPolicy;
