+++ cas-server-support-oauth/src/test/java/org/apereo/cas/support/oauth/web/OAuth20AccessTokenControllerTests.java
@@ -19,13 +19,13 @@ import org.apereo.cas.services.ServicesManager;
 import org.apereo.cas.support.oauth.OAuthConstants;
 import org.apereo.cas.support.oauth.services.OAuthRegisteredService;
 import org.apereo.cas.support.oauth.services.OAuthWebApplicationService;
-import org.apereo.cas.support.oauth.ticket.accesstoken.AccessToken;
-import org.apereo.cas.support.oauth.ticket.code.OAuthCodeFactory;
-import org.apereo.cas.support.oauth.ticket.code.DefaultOAuthCodeFactory;
-import org.apereo.cas.support.oauth.ticket.code.OAuthCode;
-import org.apereo.cas.support.oauth.ticket.refreshtoken.RefreshTokenFactory;
-import org.apereo.cas.support.oauth.ticket.refreshtoken.DefaultRefreshTokenFactory;
-import org.apereo.cas.support.oauth.ticket.refreshtoken.RefreshToken;
+import org.apereo.cas.ticket.accesstoken.AccessToken;
+import org.apereo.cas.ticket.code.OAuthCodeFactory;
+import org.apereo.cas.ticket.code.DefaultOAuthCodeFactory;
+import org.apereo.cas.ticket.code.OAuthCode;
+import org.apereo.cas.ticket.refreshtoken.RefreshTokenFactory;
+import org.apereo.cas.ticket.refreshtoken.DefaultRefreshTokenFactory;
+import org.apereo.cas.ticket.refreshtoken.RefreshToken;
 import org.apereo.cas.ticket.support.AlwaysExpiresExpirationPolicy;
 import org.junit.Before;
 import org.junit.Test;
