+++ cas-server-support-oauth/src/test/java/org/apereo/cas/support/oauth/web/OAuth20AuthorizeControllerTests.java
@@ -7,8 +7,8 @@ import org.apereo.cas.services.RegisteredService;
 import org.apereo.cas.services.ReturnAllowedAttributeReleasePolicy;
 import org.apereo.cas.support.oauth.OAuthConstants;
 import org.apereo.cas.support.oauth.services.OAuthRegisteredService;
-import org.apereo.cas.support.oauth.ticket.accesstoken.AccessToken;
-import org.apereo.cas.support.oauth.ticket.code.OAuthCode;
+import org.apereo.cas.ticket.accesstoken.AccessToken;
+import org.apereo.cas.ticket.code.OAuthCode;
 import org.junit.Test;
 import org.pac4j.cas.profile.CasProfile;
 import org.pac4j.core.context.Pac4jConstants;
