+++ cas-server-support-oauth/src/test/java/org/apereo/cas/support/oauth/web/OAuth20ProfileControllerTests.java
@@ -13,9 +13,9 @@ import org.apereo.cas.authentication.HandlerResult;
 import org.apereo.cas.authentication.TestUtils;
 import org.apereo.cas.authentication.principal.Principal;
 import org.apereo.cas.support.oauth.OAuthConstants;
-import org.apereo.cas.support.oauth.ticket.accesstoken.AccessTokenImpl;
-import org.apereo.cas.support.oauth.ticket.accesstoken.AccessTokenFactory;
-import org.apereo.cas.support.oauth.ticket.accesstoken.DefaultAccessTokenFactory;
+import org.apereo.cas.ticket.accesstoken.AccessTokenImpl;
+import org.apereo.cas.ticket.accesstoken.AccessTokenFactory;
+import org.apereo.cas.ticket.accesstoken.DefaultAccessTokenFactory;
 import org.apereo.cas.ticket.support.AlwaysExpiresExpirationPolicy;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
